+++ archive/2026-02-28_int4_id_migration/schemas/history_db/uk_aq_history_schema.sql
@@ -29,8 +29,8 @@ comment on column uk_aq_history.status_codes.created_at is
   'Creation timestamp for canonical status dictionary rows maintained for QA scripts and future validation, not ingest writes.';
 
 create table if not exists uk_aq_history.observations (
-  connector_id integer not null,
-  timeseries_id integer not null,
+  connector_id bigint not null,
+  timeseries_id bigint not null,
   observed_at timestamptz not null,
   value double precision,
   status_id smallint,
